REF: stats: rename _conting.py to _contingency.py
