+++ scipy/stats/__init__.py
@@ -10,7 +10,7 @@ from rv import *
 from morestats import *
 from kde import gaussian_kde
 import mstats
-from _conting import *
+from _contingency import *
 
 #remove vonmises_cython from __all__, I don't know why it is included
 __all__ = filter(lambda s:not (s.startswith('_') or s.endswith('cython')),dir())
