+++ scipy/stats/_contingency.py
